+++ src/main/java/vendingmachine/view/VendingMachineAccountView.java
@@ -1,9 +1,9 @@
 package vendingmachine.view;
 
-public class VendingMachineAccountInputView {
+public class VendingMachineAccountView {
 	private static final String INPUT_MESSAGE = "자판기가 보유하고 있는 금액을 입력해 주세요.";
 
-	public static void print() {
+	public static void printInputGuide() {
 		System.out.println(INPUT_MESSAGE);
 	}
 }
